Disable "Start at frame zero" as default (this option is to be removed).
Added option for adding an extra frame at the end of the animation, a copy of the first frame is used. In Blender when you are creating a loop, to avoid a duplicated frame you usually cut the last frame when it is the same as the first frame. However Urho needs this last frame, this option will add it back. Use this option only with the Timeline or Nla-Tracks. With Actions or Strips do not use it, it will cause an unwanted pause in the Urho animation.