Fix 'to_mesh'. Morph code in the trash.
If morphs have "Different vertex" errors, avoid Auto Smooth, use Face => Shade Smooth. This is because some vertices are merged in the base morph because the normals are the same, but in other morphs, the normals can diverge and this require separated vertices. A super slow pre-pass on the morphs can solve this.